Finished v3
Admin can close room(All the other players in the room are automatically removed) and start game (it starts the game for everyone in the room)
Members can leave room.
The room state(players in room. game began...) and the rooms list are updated every two seconds(only if in the relevant page)